# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cabin Reservation API
  version: 2.0.0
  description: |
    API completa para el sistema de reservas de cabañas de la cooperativa.
    
    ## Funcionalidades Principales
    
    - **Autenticación JWT**: Login, registro público, recuperación de contraseña
    - **Gestión de Reservas**: Creación, cancelación, confirmación con horarios
    - **Horarios de Check-in/Check-out**: Configurables por cabaña y reserva
    - **Gestión de Cabañas**: CRUD completo con horarios por defecto
    - **Disponibilidad**: Consulta de fechas disponibles con bloqueos obligatorios
    - **Lista de Espera**: Sistema de notificaciones y reclamo de prioridad
    - **Panel Administrativo**: Dashboard, métricas, gestión completa
    - **Precios Dinámicos**: Rangos de precios por temporada con historial
    
    ## Nuevas Funcionalidades v2.0
    
    - ✅ **Registro Público de Usuarios**: Nuevos usuarios pueden registrarse con documento válido
    - ✅ **Horarios de Check-in/Check-out**: Campos obligatorios en reservas y cabañas
    - ✅ **Validación de Documentos**: Verificación contra base de datos de asociados activos
    
  contact:
    name: Equipo de Desarrollo
    email: desarrollo@cooperativa.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
  - url: https://api.cooperativa.com
    description: Servidor de producción
security:
  - bearerAuth: []
paths:
...
  /api/admin/documents/number/{documentNumber}:
    get:
      tags:
        - 👑 Admin - Documents
      summary: Obtener documento por número
      description: Retorna detalles de un documento por su número
      parameters:
        - name: documentNumber
          in: path
          required: true
          description: Número del documento
          schema:
            type: string
            example: "99999999"
      responses:
        '200':
          description: Documento obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentNumberResponse'
        '404':
          description: Documento no encontrado
  /api/admin/documents:
    post:
      tags:
        - 👑 Admin - Documents
      summary: Crear nuevo documento
      description: Crea un nuevo número de documento en el sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentNumberRequest'
            examples:
              new_document:
                summary: Crear documento de prueba
                value:
                  documentNumber: "88888888"
                  status: "ACTIVE"
      responses:
        '201':
          description: Documento creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentNumberResponse'
        '400':
          description: Documento ya existe o solicitud inválida
  /api/admin/documents/{id}/activate:
    put:
      tags:
        - 👑 Admin - Documents
      summary: Activar documento
      description: Activa un documento cambiando su estado a ACTIVE
      parameters:
        - name: id
          in: path
          required: true
          description: ID del documento
          schema:
            type: integer
            example: 7
      responses:
        '200':
          description: Documento activado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Documento activado exitosamente"
  /api/admin/documents/{id}/deactivate:
    put:
      tags:
        - 👑 Admin - Documents
      summary: Desactivar documento
      description: Deshabilita un documento cambiando su estado a DISABLED
      parameters:
        - name: id
          in: path
          required: true
          description: ID del documento
          schema:
            type: integer
            example: 7
      responses:
        '200':
          description: Documento desactivado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Documento desactivado exitosamente"
  /api/admin/documents/{id}:
    delete:
      tags:
        - 👑 Admin - Documents
      summary: Eliminar documento
      description: Elimina permanentemente un documento del sistema
      parameters:
        - name: id
          in: path
          required: true
          description: ID del documento
          schema:
            type: integer
            example: 7
      responses:
        '200':
          description: Documento eliminado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Documento eliminado exitosamente"
        '404':
          description: Documento no encontrado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint de login o registro
  schemas:
...
    DocumentNumberResponse:
      type: object
      properties:
        id:
          type: integer
          description: ID único del documento
          example: 7
        documentNumber:
          type: string
          description: Número de documento
          example: "99999999"
        status:
          type: string
          enum: [ACTIVE, DISABLED]
          description: Estado del documento
          example: "ACTIVE"
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación
          example: "2024-01-01T10:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Fecha de última actualización
          example: "2024-01-01T10:00:00.000Z"
    CreateDocumentNumberRequest:
      type: object
      required:
        - documentNumber
      properties:
        documentNumber:
          type: string
          description: Número de documento
          minLength: 7
          maxLength: 20
          pattern: '^\d+$'
          example: "88888888"
        status:
          type: string
          enum: [ACTIVE, DISABLED]
          description: Estado del documento
          default: ACTIVE
          example: "ACTIVE"
...